Refactor to get each reviewer's info from API instead of from Firestore
